api-service/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
httpx  # For TestClient in testing
newsapi-python
pinecone-client 